feat: update ingredient
